adaptive_gamification_rl (Python)/requirements.txt
torch
numpy
matplotlib
fastapi
pyyaml